+++ python/Web_scraping/spreadsheet/tests/test_spreadsheet.py
@@ -3,10 +3,10 @@
 """
 検証コード
 """
-
+import os
 import unittest
-from tenki import *
 from const import *
+from tenki.tenki.tenki import *
 
 
 class TestSpreadsheet(unittest.TestCase):
@@ -15,13 +15,13 @@ class TestSpreadsheet(unittest.TestCase):
 
     def setUp(self):
         print("setUp")
-        json_keyfile_name = 'C:\\Git\\igapon50\\traning\\python\\Web_scraping\\tenki-347610-1bc0fec79f90.json'
-        workbook_name = '天気予報'
-        worksheet_name = '七尾市和倉町data'
-        self.spreadsheet = Spreadsheet(json_keyfile_name,
-                                  workbook_name,
-                                  worksheet_name,
-                                  )
+        self.json_keyfile_name = (os.path.dirname(__file__) + r"\..\..\json\tenki-347610-1bc0fec79f90.json")
+        self.workbook_name = '天気予報'
+        self.worksheet_name = '七尾市和倉町data'
+        self.spreadsheet = Spreadsheet(self.json_keyfile_name,
+                                       self.workbook_name,
+                                       self.worksheet_name,
+                                       )
 
     def tearDown(self):
         print("tearDown")
@@ -51,9 +51,7 @@ class TestSpreadsheet(unittest.TestCase):
         :return:
         """
         spreadsheet3 = Spreadsheet()
-        spreadsheet3.load_text(RESULT_FILE_PATH + '3.txt')
-        spreadsheet3.save_text(RESULT_FILE_PATH + '4.txt')
-
+        spreadsheet3.load_text(RESULT_FILE_PATH + '1.txt')
         spreadsheet3.write_list_columns([100, 200, 300, 50], (1, 1))
         spreadsheet3.write_list_columns([99, 98, 97, 96], (1, 3))
         spreadsheet3.write_list_rows([1, 2, 3, 4], (5, 5))
@@ -86,77 +84,6 @@ class TestSpreadsheet(unittest.TestCase):
                                 "4m/s", "2m/s", "2m/s", "2m/s", "2m/s", "2m/s", "3m/s"]}
         spreadsheet3.write_dict_columns(tenki, (10, 10))
         spreadsheet3.write_dict_rows(tenki, (10, 17))
-
-    def test_tenki01(self):
-        css_root = "dd.forecast10days-actab"
-        css_selectors = {"days_item": "div.days",
-                         "time_item": "dd.time-item > span",
-                         "forecast_item": "dd.forecast-item > p > img",
-                         "prob_precip_item": "dd.prob-precip-item > span > span",
-                         "precip_item": "dd.precip-item > span > span",
-                         "temp_item": "dd.temp-item > script",
-                         "wind_item_blow": "dd.wind-item > p > img",
-                         "wind_item_speed": "dd.wind-item > p > span",
-                         }
-        attrs = {"days_item": "",
-                 "time_item": "",
-                 "forecast_item": "alt",
-                 "prob_precip_item": "",
-                 "precip_item": "",
-                 "temp_item": "",
-                 "wind_item_blow": "alt",
-                 "wind_item_speed": "",
-                 }
-        tenki = Tenki("https://tenki.jp/forecast/4/20/5620/17202/10days.html",
-                      css_root,
-                      css_selectors,
-                      attrs,
-                      )
-        tenki.save_text(RESULT_FILE_PATH + '1.txt')
-        # 値オブジェクトを生成
-        value_objects = tenki.get_value_objects()
-        # 値オブジェクトでインスタンス作成
-        tenki2 = Tenki(value_objects)
-        # 保存や読込を繰り返す
-        tenki2.save_text(RESULT_FILE_PATH + '2.txt')
-        tenki2.load_text(RESULT_FILE_PATH + '2.txt')
-        tenki2.save_text(RESULT_FILE_PATH + '3.txt')
-        tenki3 = Tenki()
-        tenki3.load_text(RESULT_FILE_PATH + '3.txt')
-        tenki3.save_text(RESULT_FILE_PATH + '4.txt')
-
-    def test_tenki02(self):
-        tenki = Tenki()
-        tenki.load_text(RESULT_FILE_PATH + '1.txt')
-        json_keyfile_name = 'C:\\Git\\igapon50\\traning\\python\\Web_scraping\\tenki-347610-1bc0fec79f90.json'
-        workbook_name = '天気予報'
-        worksheet_name = '七尾市和倉町data'
-        spreadsheet = Spreadsheet(json_keyfile_name,
-                                  workbook_name,
-                                  worksheet_name,
-                                  )
-        spreadsheet.save_text(RESULT_FILE_PATH + '2.txt')
-        spreadsheet.write_dict_columns(tenki.get_result_forecasts(), (1, 1))
-        num = len(tenki.get_result_forecasts())
-        spreadsheet.write_dict_columns(tenki.get_result_counters(), (1, 1 + num))
-        worksheet_name = '七尾市和倉町conv'
-        spreadsheet = Spreadsheet(json_keyfile_name,
-                                  workbook_name,
-                                  worksheet_name,
-                                  )
-        spreadsheet.save_text(RESULT_FILE_PATH + '3.txt')
-        tenki.special_func_temp()
-        spreadsheet.write_dict_columns(tenki.get_result_forecasts(), (1, 1))
-        num = len(tenki.get_result_forecasts())
-        spreadsheet.write_dict_columns(tenki.get_result_counters(), (1, 1 + num))
-        spreadsheet.save_text(RESULT_FILE_PATH + '4.txt')
-        tenki.save_text(RESULT_FILE_PATH + '5.txt')
-
-    def test_tenki03(self):
-        tenki = Tenki()
-        tenki.load_text(RESULT_FILE_PATH + '1.txt')
-        tenki.special_func_temp()
-        tenki.create_LINE_BOT_TOBA_format()
 
 
 if __name__ == "__main__":
