--- python/Web_scraping/test.bat
@@ -1,2 +1,0 @@
-@echo off
-python test.py %1 %2
